Restructure testing and directive block parsing (#59)
In this PR I have:

1. Restructured the test folder and added an `AstRenderer` and lower level tests for the source text to Markdown (mistletoe) AST.
2. Added `docs/develop/test_infrastructure.md` to explain the testing inrastructure.
3. Restructured the directive block parsing, to make it more modular/understandable and improved error checking and reporting.

fixes #54, fixes #55, fixes #51
