+++ tests/test_commonmark/test_commonmark.py
@@ -1,3 +1,6 @@
+"""In this module tests are run against the full test set,
+provided by https://github.com/commonmark/CommonMark.git.
+"""
 import json
 import os
