+++ tests/test_renderers/test_docutils/test_cross_referencing.xml
@@ -1,0 +1,18 @@
+<document source="">
+    <target ids="target" names="target">
+    <section ids="title" names="title">
+        <title>
+            Title
+        <paragraph>
+            <pending_xref refdomain="True" refexplicit="True" reftarget="target" reftype="any" refwarn="True">
+                <TextElement classes="xref any">
+                    alt1
+        <paragraph>
+            <pending_xref refdomain="True" refexplicit="False" reftarget="target2" reftype="any" refwarn="True">
+                <TextElement classes="xref any">
+        <paragraph>
+            <reference refuri="https://www.google.com">
+                alt2
+        <paragraph>
+            <reference refuri="#target3">
+                alt3
