+++ tests/test_renderers/test_roles_directives/test_directive_options_error_colon_style_text1_.xml
@@ -1,0 +1,13 @@
+<document source="">
+    <system_message level="3" line="0" source="" type="ERROR">
+        <paragraph>
+            Directive 'restructuredtext-test-directive' options:
+            Invalid YAML: mapping values are not allowed here
+              in "<unicode string>", line 2, column 8:
+                option2: b
+                       ^
+        <literal_block xml:space="preserve">
+            :option1
+            :option2: b
+            
+            content
