+++ tests/test_syntax/test_ast/test_comment_basic_strings0_.yml
@@ -1,0 +1,5 @@
+children:
+- content: comment
+  type: LineComment
+footnotes: {}
+type: Document
