+++ tests/test_syntax/test_ast/test_comment_escaped_strings3_.yml
@@ -1,0 +1,14 @@
+children:
+- children:
+  - children:
+    - content: '%'
+      type: RawText
+    type: EscapeSequence
+  - content: ' comment'
+    type: RawText
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
