+++ tests/test_syntax/test_ast/test_comment_follows_list_strings5_.yml
@@ -1,0 +1,21 @@
+children:
+- children:
+  - children:
+    - children:
+      - content: item
+        type: RawText
+      range:
+      - 1
+      - 1
+      type: Paragraph
+    - content: comment
+      type: LineComment
+    leader: '-'
+    loose: false
+    prepend: 2
+    type: ListItem
+  loose: false
+  start: null
+  type: List
+footnotes: {}
+type: Document
