+++ tests/test_syntax/test_ast/test_comment_indent_2_strings1_.yml
@@ -1,0 +1,5 @@
+children:
+- content: comment
+  type: LineComment
+footnotes: {}
+type: Document
