+++ tests/test_syntax/test_ast/test_comment_indent_4_strings2_.yml
@@ -1,0 +1,13 @@
+children:
+- children:
+  - content: '% comment
+
+      '
+    type: RawText
+  language: ''
+  range:
+  - 0
+  - 1
+  type: BlockCode
+footnotes: {}
+type: Document
