+++ tests/test_syntax/test_ast/test_comment_inline_strings4_.yml
@@ -1,0 +1,10 @@
+children:
+- children:
+  - content: a % comment
+    type: RawText
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
