+++ tests/test_syntax/test_ast/test_front_matter_basic_strings0_.yml
@@ -1,0 +1,11 @@
+children:
+- children: []
+  content: 'a: b
+
+    '
+  range:
+  - 0
+  - 3
+  type: FrontMatter
+footnotes: {}
+type: Document
