+++ tests/test_syntax/test_ast/test_math_basic_strings0_.yml
@@ -1,0 +1,10 @@
+children:
+- children:
+  - content: $a$
+    type: Math
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
