+++ tests/test_syntax/test_ast/test_math_contains_special_chars_strings1_.yml
@@ -1,0 +1,10 @@
+children:
+- children:
+  - content: $a`{_*-%$
+    type: Math
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
