+++ tests/test_syntax/test_ast/test_math_escaped_opening_strings4_.yml
@@ -1,0 +1,16 @@
+children:
+- children:
+  - children:
+    - content: $
+      type: RawText
+    type: EscapeSequence
+  - content: 'a '
+    type: RawText
+  - content: $b$
+    type: Math
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
