+++ tests/test_syntax/test_ast/test_math_issue_51_strings8_.yml
@@ -1,0 +1,18 @@
+children:
+- children:
+  - children:
+    - content: $x_{hey}=it+is^{math}$
+      type: RawText
+    type: InlineCode
+  - content: ' renders as '
+    type: RawText
+  - content: $x_{hey}=it+is^{math}$
+    type: Math
+  - content: .
+    type: RawText
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
