+++ tests/test_syntax/test_ast/test_math_multiple_strings3_.yml
@@ -1,0 +1,14 @@
+children:
+- children:
+  - content: $a$
+    type: Math
+  - content: ' '
+    type: RawText
+  - content: $b$
+    type: Math
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
