+++ tests/test_syntax/test_ast/test_math_no_closing_strings5_.yml
@@ -1,0 +1,10 @@
+children:
+- children:
+  - content: $a
+    type: RawText
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
