+++ tests/test_syntax/test_ast/test_math_preceding_special_chars_strings2_.yml
@@ -1,0 +1,12 @@
+children:
+- children:
+  - content: '{_*-%`'
+    type: RawText
+  - content: $a$
+    type: Math
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
