+++ tests/test_syntax/test_ast/test_target_escaped_strings3_.yml
@@ -1,0 +1,14 @@
+children:
+- children:
+  - children:
+    - content: (
+      type: RawText
+    type: EscapeSequence
+  - content: target)=
+    type: RawText
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
