+++ tests/test_syntax/test_ast/test_target_external_emphasis_strings6_.yml
@@ -1,0 +1,15 @@
+children:
+- children:
+  - children:
+    - children:
+      - content: target
+        type: RawText
+      target: target
+      type: Target
+    type: Emphasis
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
