+++ tests/test_syntax/test_ast/test_target_indent_2_strings1_.yml
@@ -1,0 +1,13 @@
+children:
+- children:
+  - children:
+    - content: target
+      type: RawText
+    target: target
+    type: Target
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
