+++ tests/test_syntax/test_ast/test_target_inline_strings4_.yml
@@ -1,0 +1,15 @@
+children:
+- children:
+  - content: 'a '
+    type: RawText
+  - children:
+    - content: target
+      type: RawText
+    target: target
+    type: Target
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
