+++ tests/test_syntax/test_ast/test_target_internal_emphasis_strings5_.yml
@@ -1,0 +1,13 @@
+children:
+- children:
+  - children:
+    - content: '*target*'
+      type: RawText
+    target: '*target*'
+    type: Target
+  range:
+  - 1
+  - 1
+  type: Paragraph
+footnotes: {}
+type: Document
